#65 rename: records 폴더 내 파일 dashboard 폴더로 이동해 url 변경
