+++ app/dashboard/records.module.scss
@@ -11,18 +11,18 @@
 }
 @font-face {
   font-family: "Arimo_BLD";
-  src: url("../../../public/Arimo-Bold.ttf");
+  src: url("../../public/Arimo-Bold.ttf");
   font-weight: 700;
 }
 
 @font-face {
   font-family: "Arimo_RGR";
-  src: url("../../../public/Arimo-Regular.ttf");
+  src: url("../../public/Arimo-Regular.ttf");
 }
 
 @font-face {
   font-family: "Arimo_MED";
-  src: url("../../../public/Arimo-Medium.ttf");
+  src: url("../../public/Arimo-Medium.ttf");
 }
 
 $black: #383838;
